AnimationPlayBack behaviour change and ownership
Now when the animationPlayBack is played once or looped, it requires an animation to be passed so that it can initalise itself to handle the frame of the specified animation. Scene no longer has a dictonary of AnimationPlayBack controllers as they are nowed owned by each model. AnimationPlayBack controllers are no longer binded to a single animation upon initalisation. the animation can be changed easlily by passing in the animation to the play or loop functions